// src/app/enum/settings.enum.ts
// enums/settings.enum.ts

export enum WhoCanSeeLocation {
  FRIENDS_ONLY = "friends_only",
  EVERYONE = "everyone",
  NOBODY = "nobody",
}

export enum WhoCanSeeTrip {
  FRIENDS_ONLY = "friends_only",
  EVERYONE = "everyone",
  NOBODY = "nobody",
}

export enum WhoCanSendFriendRequest {
  EVERYONE = "everyone",
  NOBODY = "nobody",
}